style(msvc): fix C4244/C4267 in XmlToLua/WheelBase/StatsManager/ScreenOptions/Course/ActorMultiTexture (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226.

- XmlToLua.cpp (6): RString::Left/Right fed size_t arithmetic at 5
  call sites parsing field names and file paths
  (add_extension_to_relative_path..., store_cmd, store_field,
  load_frames_from_file x2); a trailing-.lua filename Left() call.
- WheelBase.cpp (6): a size_t item count into int; five
  wrap(int&, size_t) sites across ChangeMusic/RebuildWheelItems/
  FirstVisibleIndex.
- StatsManager.cpp (6): a size_t song count into unsigned; a
  std::trunc() result and two size_t accumulators feeding profile
  gameplay-seconds/songs-played counters; a size_t-minus-int index; a
  lua_pushnumber(size_t) accessor.
- ScreenOptions.cpp (6): a MoveRowAbsolute call's size_t-minus-1 arg;
  the same 'const int iNumChoices = ...m_vsChoices.size()' pattern at
  2 call sites (4 warnings, both int/const int variants); a
  wrap(int&, size_t) site.
- Course.cpp (6): a size_t mod-change count into int; two
  RandomInt(size_t) calls choosing steps; two identical
  std::floor((iMinDist + iMaxDist) / 2) sites promoting an all-int
  expression to double then narrowing back, in two near-duplicate
  meter-balancing code paths; a lua_pushnumber(size_t) accessor.
- ActorMultiTexture.cpp (6): two int-into-float texture-size members;
  two size_t-into-int texture-unit-count returns; two
  enum_add2(TextureUnit_1, size_t) calls needing an int second arg.

+++ src/ScreenOptions.cpp
@@ -970,7 +970,7 @@ void ScreenOptions::ProcessMenuStart( const InputEventPlus &input )
 		case NAV_FIVE_KEY:
 			/* Jump to the exit row.  (If everyone's already on the exit row, then
 			 * we'll have already gone to the next screen above.) */
-			if( MoveRowAbsolute(pn, m_pRows.size()-1) )
+			if( MoveRowAbsolute(pn, static_cast<int>(m_pRows.size())-1) )
 				m_SoundNextRow.Play(true);
 
 			break;
@@ -1035,7 +1035,7 @@ void ScreenOptions::ChangeValueInRowAbsolute( int iRow, PlayerNumber pn, int iCh
 
 	OptionRow &row = *m_pRows[iRow];
 
-	const int iNumChoices = row.GetRowDef().m_vsChoices.size();
+	const int iNumChoices = static_cast<int>(row.GetRowDef().m_vsChoices.size());
 	ASSERT( iNumChoices >= 0 && iChoiceIndex < iNumChoices );
 
 	int iCurrentChoiceWithFocus = row.GetChoiceInRowWithFocus(pn);
@@ -1056,7 +1056,7 @@ void ScreenOptions::ChangeValueInRowRelative( int iRow, PlayerNumber pn, int iDe
 
 	OptionRow &row = *m_pRows[iRow];
 
-	const int iNumChoices = row.GetRowDef().m_vsChoices.size();
+	const int iNumChoices = static_cast<int>(row.GetRowDef().m_vsChoices.size());
 
 	if( m_OptionsNavigation == NAV_THREE_KEY_MENU && iNumChoices <= 1 )	// 1 or 0
 	{
@@ -1172,7 +1172,7 @@ bool ScreenOptions::MoveRowRelative( PlayerNumber pn, int iDir, bool bRepeat )
 	{
 		int iDelta = r*iDir;
 		iDest = m_iCurrentRow[pn] + iDelta;
-		wrap( iDest, m_pRows.size() );
+		wrap( iDest, static_cast<int>(m_pRows.size()) );
 
 		OptionRow &row = *m_pRows[iDest];
 		if( row.GetRowDef().IsEnabledForPlayer(pn) )
